refactor: organize library feature structure
