+++ src/features/library/types.ts
@@ -2,6 +2,16 @@ import type { Bracket } from "../../data/mockRuns";
 import type { HomeFilterState } from "../home/types";
 
 export type NumericRange = { min: number; max: number };
+export type LibraryFilterKey = "character" | "build" | "category" | "tag" | "search";
+export type SelectableFilterKey = Exclude<LibraryFilterKey, "search">;
+export type SearchFilterKey = Extract<LibraryFilterKey, "search">;
+export type LibraryFilterPanel = { key: LibraryFilterKey; label: string; icon: string };
+export type SelectableFilterPanel = LibraryFilterPanel & { key: SelectableFilterKey };
+export type SearchFilterPanel = LibraryFilterPanel & { key: SearchFilterKey };
+export type CharacterSummaryGroup = "partyCharacters" | "mainAttackers";
+export type CharacterSummaryTarget = "include" | "exclude";
+export type WeaponSummaryTarget = "include" | "exclude";
+export type LibraryBuildFilterTab = "cost" | "weapon";
 
 export type LibraryCategoryFilterState = {
   ruleset: string;
